# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: A Dynamic DNS REST Service.
  description: The routes of the dynamic DNS API
  contact:
    name: Thilo Schwarz
    url: https://github.com/th-schwarz/DynDRest
    email: osp (at) thischwa.codes
  license:
    name: MIT Licence
    url: https://github.com/th-schwarz/DynDRest/blob/develop/LICENSE
  version: "0.4"
...
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /admin/zones/{zoneName}/{ns}:
    post:
      tags:
      - admin-controller
      summary: Adds a zone with the specified name and name server.
      operationId: addZone
      parameters:
      - name: zoneName
        in: path
        description: "The name of the zone, must be a valid domain name."
        required: true
        schema:
          type: string
          description: "The name of the zone, must be a valid domain name."
          example: domain.com
        example: domain.com
      - name: ns
        in: path
        description: "The name server of the zone, must be a valid domain name."
        required: true
        schema:
          type: string
          description: "The name server of the zone, must be a valid domain name."
          example: ns1.domain.info
        example: ns1.domain.info
      responses:
        "200":
          description: The desired Zone was successfully created.
        "409":
          description: A zone with the specified name already exists.
  /admin/zones/{zoneName}/hosts/{host}:
    post:
      tags:
      - admin-controller
      summary: Adds a host with the specified host for the desired zone into the local
        storage.
      operationId: addHost
      parameters:
      - name: zoneName
        in: path
        description: The name of the zone for which the host is to be created.
        required: true
        schema:
          type: string
          description: The name of the zone for which the host is to be created.
          example: domain.com
        example: domain.com
      - name: host
        in: path
        description: "The name of the host to create, it's just the sub-domain part."
        required: true
        schema:
          type: string
          description: "The name of the host to create, it's just the sub-domain part."
          example: host1
        example: host1
      - name: apiToken
        in: query
        description: The 'api-token' to authenticate IP changes of the created host.
          Should be a strong one!
        required: true
        schema:
          type: string
          description: The 'api-token' to authenticate IP changes of the created host.
            Should be a strong one!
      responses:
        "200":
          description: The desired host was successfully created.
        "404":
          description: The zone doesn't exists.
        "409":
          description: The desired host already exists.
        "500":
          description: If the operation has failed.
...
  /admin/zones/{zoneName}/hosts:
    get:
      tags:
      - admin-controller
      summary: Returns a list of hosts of the desired zone.
      operationId: listHostsOfZone
      parameters:
      - name: zoneName
        in: path
        description: The name of the zone for which the hosts are to be listed.
        required: true
        schema:
          type: string
          description: The name of the zone for which the hosts are to be listed.
          example: domain.com
        example: domain.com
      responses:
        "200":
          description: The hosts of the desired zones fetched successful.
          content:
            application/json:
              example:
              - name: master
                apiToken: secureApiToken
                zone: mydomain.net
                ns: a1.nameserver.net
                fullHost: master.mydomain.net
                changed: 2024-03-11T09:07:59.037688
  /admin/zones/{name}:
    delete:
      tags:
      - admin-controller
      summary: Deletes the zone with the specified name.
      operationId: deleteZone
      parameters:
      - name: name
        in: path
        description: "The name of the zone, must be a valid domain name."
        required: true
        schema:
          type: string
          description: "The name of the zone, must be a valid domain name."
          example: domain.com
        example: domain.com
      responses:
        "200":
          description: The Zone deleted successful.
        "404":
          description: The zone doesn't exists.
  /admin/zones/hosts/{host}:
    delete:
      tags:
      - admin-controller
      summary: Deletes the host with the specified host name.
      operationId: deleteHost
      parameters:
      - name: host
        in: path
        description: The full name of the host.
        required: true
        schema:
          type: string
          description: The full name of the host.
          example: mydyndns.domain.com
        example: mydyndns.domain.com
      responses:
        "200":
          description: The desired host was successfully deleted.
        "404":
          description: The desired host doesn't exists.
        "500":
          description: If the operation has failed.
...