# src/main/java/codes/thischwa/dyndrest/server/controller/AdminController.java
package codes.thischwa.dyndrest.server.controller;

import codes.thischwa.dyndrest.model.HostEnriched;
import codes.thischwa.dyndrest.model.Zone;
import codes.thischwa.dyndrest.provider.Provider;
import codes.thischwa.dyndrest.provider.ProviderException;
import codes.thischwa.dyndrest.service.HostZoneService;
import codes.thischwa.dyndrest.util.DomainNameValidator;
import java.util.List;
import java.util.Optional;
import lombok.extern.slf4j.Slf4j;
import org.springframework.dao.DuplicateKeyException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.server.ResponseStatusException;

/** The AdminController provides methods for managing zones and hosts in the system. */
@Controller
@Slf4j
public class AdminController implements AdminRoutes {

  private final DomainNameValidator domainNameValidator = new DomainNameValidator();
  private final HostZoneService hostZoneService;
  private final Provider provider;

  /**
   * Creates a new instance of the AdminController class.
   *
   * @param hostZoneService The host zone service.
   * @param provider The provider.
   */
  public AdminController(HostZoneService hostZoneService, Provider provider) {
    this.hostZoneService = hostZoneService;
    this.provider = provider;
  }

  @Override
  public ResponseEntity<Void> addZone(String zoneName, String ns) {
    log.debug("entered #addZone: name={}, ns={}", zoneName, ns);
    if (!domainNameValidator.isValidDomainNames(zoneName, ns)) {
      log.error("Domain name ({}) or nameserver ({}) is malformed.", zoneName, ns);
      throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
    }
    try {
      hostZoneService.addZone(zoneName, ns);
    } catch (DuplicateKeyException e) {
      log.error("Zone with name {} already exists.", zoneName);
      throw new ResponseStatusException(HttpStatus.CONFLICT);
    } catch (Exception e) {
      log.error("Failed to add zone: name={}, ns={}", zoneName, ns, e);
      throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
    }
    return ResponseEntity.ok().build();
  }

  @Override
  public ResponseEntity<Void> deleteZone(String name) {
    log.debug("entered #deleteZone: name={}", name);
    Zone zone = hostZoneService.getZone(name);
    if (zone == null) {
      log.error("Zone with name {} not found.", name);
      throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
    }
    hostZoneService.deleteZone(zone);
    return ResponseEntity.ok().build();
  }

  @Override
  public ResponseEntity<List<Zone>> listZones() {
    log.debug("entered #listZones.");
    List<Zone> zones = hostZoneService.getAllZones();
    return ResponseEntity.ok(zones);
  }

  @Override
  public ResponseEntity<List<HostEnriched>> listHostsOfZone(String zoneName) {
    log.debug("entered #listHostsOfZone.");
    Optional<List<HostEnriched>> hosts = hostZoneService.findHostsOfZone(zoneName);
    return hosts.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
  }

  @Override
  public ResponseEntity<Void> addHost(
      String zoneName, String host, String apiToken) {
    log.debug("entered #addHost: name={}, host={}", zoneName, host);
    Zone zone = hostZoneService.getZone(zoneName);
    if (zone == null) {
      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
    }
    Optional<List<HostEnriched>> fullHosts = hostZoneService.findHostsOfZone(zoneName);
    if (fullHosts.isPresent() && fullHosts.get().stream().anyMatch(h -> h.getSld().equals(host))) {
      log.error("Host {} already exists.", host);
      throw new ResponseStatusException(HttpStatus.CONFLICT);
    }

    // add host into the database
    hostZoneService.addHost(zone, host, apiToken);
    return ResponseEntity.ok().build();
  }

  @Override
  public ResponseEntity<Void> deleteHost(String host) {
    log.debug("entered #deleteHost: fullHost={}", host);
    Optional<HostEnriched> optionalFullHost = hostZoneService.getHost(host);
    if (optionalFullHost.isEmpty()) {
      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
    }
    HostEnriched hostEnriched = optionalFullHost.get();
    try {
      provider.removeHostIpSettings(hostEnriched.getFullHost());
    } catch (ProviderException e) {
      log.error("Failed to remove host: fullHost=" + hostEnriched.getFullHost(), e);
      throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
    }
    hostZoneService.deleteHost(hostEnriched);
    return ResponseEntity.ok().build();
  }
}